@@ -3,9 +3,9 @@
   "displayName": "Color Pick",
   "description": "Open a color picker from any file. Works in JSX. Supports hex or rgba. No dependencies.",
   "icon": "pick.png",
-  "version": "0.6.0",
+  "version": "0.7.0",
   "engines": {
-    "vscode": "^1.23.0"
+    "vscode": "^1.85.0"
   },
   "publisher": "adam-watters",
   "author": {
@@ -16,61 +16,48 @@
     "type": "git",
     "url": "https://github.com/adamwatters/vscode-color-pick"
   },
-  "activationEvents": [
-    "onCommand:pick-color"
-  ],
-  "main": "./build/ext-src/extension.js",
+  "main": "./build/extension.js",
   "contributes": {
     "commands": [
       {
         "command": "pick-color",
         "title": "Pick Color"
       }
-    ],
-    "configuration": {
-      "type": "object",
-      "title": "Color Pick configuration",
-      "properties": {
-        "color-pick.analytics": {
-          "type": "boolean",
-          "default": false,
-          "description": "Allow Color Pick to collect color search events and color selection events."
-        }
-      }
-    }
+    ]
   },
   "license": "MIT",
   "dependencies": {
-    "@types/color-convert": "^1.9.0",
-    "@types/lodash": "^4.14.133",
-    "@types/react-color": "^3.0.0",
     "color-convert": "^2.0.0",
-    "color-name-list": "^4.7.1",
-    "fuse.js": "^3.4.4",
-    "lodash": "^4.17.11",
-    "react": "^16.8.6",
-    "react-color": "^2.17.3",
-    "react-dom": "^16.8.6",
-    "react-scripts-ts": "^3.1.0",
-    "vscode-extension-telemetry": "^0.1.1"
+    "color-name-list": "^10.0.0",
+    "fuse.js": "^7.0.0",
+    "react": "^18.3.0",
+    "react-color": "^2.19.3",
+    "react-dom": "^18.3.0"
   },
   "scripts": {
-    "vscode:prepublish": "react-scripts-ts build && tsc -p tsconfig.extension.json",
-    "postinstall": "node ./node_modules/vscode/bin/install",
-    "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build && tsc -p tsconfig.extension.json",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject"
+    "build": "vite build && esbuild ext-src/extension.ts --bundle --outfile=build/extension.js --external:vscode --format=cjs --platform=node --sourcemap",
+    "build:webview": "vite build",
+    "build:extension": "esbuild ext-src/extension.ts --bundle --outfile=build/extension.js --external:vscode --format=cjs --platform=node --sourcemap",
+    "vscode:prepublish": "yarn build",
+    "dev:webview": "vite",
+    "lint": "eslint src ext-src"
   },
   "devDependencies": {
-    "@types/jest": "^22.2.3",
-    "@types/node": "^10.1.2",
-    "@types/react": "^16.8.18",
-    "@types/react-dom": "^16.0.5",
-    "@types/react-test-renderer": "^16.8.1",
-    "all-contributors-cli": "^6.6.0",
-    "react-test-renderer": "^16.8.6",
-    "typescript": "^2.8.3",
-    "vscode": "^1.1.17"
+    "@types/node": "^20.0.0",
+    "@types/react": "^18.3.0",
+    "@types/react-color": "^3.0.0",
+    "@types/react-dom": "^18.3.0",
+    "@types/vscode": "^1.85.0",
+    "@types/color-convert": "^2.0.0",
+    "@vitejs/plugin-react": "^4.3.0",
+    "esbuild": "^0.24.0",
+    "eslint": "^9.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.0.0",
+    "@typescript-eslint/parser": "^8.0.0",
+    "eslint-plugin-react": "^7.37.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "typescript": "^5.7.0",
+    "typescript-eslint": "^8.0.0",
+    "vite": "^6.0.0"
   }
 }
