@@ -1,43 +1,16 @@
+import * as fs from "fs";
 import * as path from "path";
 import * as vscode from "vscode";
-import TelemetryReporter from "vscode-extension-telemetry";
-const environment =
-  process.env.VSCODE_DEBUG_MODE === "true" ? "development" : "production";
-let reporter: TelemetryReporter;
 
 export function activate(context: vscode.ExtensionContext) {
-  const extensionId = "adam-watters.vscode-color-pick";
-  const key = "4a812e02-fb45-447f-a2bc-42bf98535773";
-  const extension = vscode.extensions.getExtension(extensionId) || {
-    packageJSON: "undefined"
-  };
-  reporter = new TelemetryReporter(
-    extensionId,
-    extension.packageJSON.version,
-    key
-  );
   context.subscriptions.push(
     vscode.commands.registerCommand("pick-color", () => {
-      ReactPanel.createOrShow(
-        context.extensionPath,
-        context.globalState,
-        reporter
-      );
+      ReactPanel.createOrShow(context.extensionPath, context.globalState);
     })
   );
 }
 
-export function deactivate() {
-  reporter.dispose();
-}
-
-const shouldUseTelemetry = (): boolean => {
-  return (
-    process.env.VSCODE_DEBUG_MODE === "true" ||
-    vscode.workspace.getConfiguration().get("color-pick.analytics") ||
-    false
-  );
-};
+export function deactivate() {}
 
 const defaultState = {
   color: {
@@ -46,19 +19,13 @@ const defaultState = {
       a: 1,
       b: 65,
       g: 255,
-      r: 0
-    }
+      r: 0,
+    },
   },
-  mode: "rgba"
+  mode: "rgba",
 };
 
-/**
- * Manages react webview panels
- */
 class ReactPanel {
-  /**
-   * Track the currently panel. Only allow a single panel to exist at a time.
-   */
   public static currentPanel: ReactPanel | undefined;
 
   private static readonly viewType = "color-pick";
@@ -70,15 +37,13 @@ class ReactPanel {
 
   public static async createOrShow(
     extensionPath: string,
-    globalState: vscode.Memento,
-    reporter: TelemetryReporter
+    globalState: vscode.Memento
   ) {
     const column = vscode.window.activeTextEditor
       ? vscode.window.activeTextEditor.viewColumn
       : undefined;
     const initialState = await globalState.get("app");
-    // If we already have a panel, show it.
-    // Otherwise, create a new panel.
+
     if (ReactPanel.currentPanel) {
       ReactPanel.currentPanel._panel.reveal(column);
     } else {
@@ -86,8 +51,7 @@ class ReactPanel {
         extensionPath,
         globalState,
         initialState || defaultState,
-        column || vscode.ViewColumn.One,
-        reporter
+        column || vscode.ViewColumn.One
       );
     }
   }
@@ -96,69 +60,49 @@ class ReactPanel {
     extensionPath: string,
     globalState: vscode.Memento,
     initialState: object,
-    column: vscode.ViewColumn,
-    reporter: TelemetryReporter
+    column: vscode.ViewColumn
   ) {
     this._extensionPath = extensionPath;
     this._globalState = globalState;
-    // Create and show a new webview panel
+
     this._panel = vscode.window.createWebviewPanel(
       ReactPanel.viewType,
       "Color Pick",
       column,
       {
-        // Enable javascript in the webview
         enableScripts: true,
-
-        // And restric the webview to only loading content from our extension's `media` directory.
         localResourceRoots: [
-          vscode.Uri.file(path.join(this._extensionPath, "build"))
-        ]
+          vscode.Uri.file(path.join(this._extensionPath, "build", "webview")),
+        ],
       }
     );
 
-    // Set the webview's initial html content
     this._panel.webview.html = this._getHtmlForWebview(initialState);
 
-    // Listen for when the panel is disposed
-    // This happens when the user closes the panel or when the panel is closed programatically
     this._panel.onDidDispose(() => this.dispose(), null, this._disposables);
 
-    // Handle messages from the webview
     this._panel.webview.onDidReceiveMessage(
-      message => {
+      (message) => {
         switch (message.command) {
-          case "search":
-            const { searchString } = message;
-            shouldUseTelemetry() &&
-              reporter.sendTelemetryEvent("search", {
-                searchString,
-                environment
-              });
-          case "colorChanged":
+          case "colorChanged": {
             const {
               mode,
               color: {
                 rgb: { r, g, b, a },
-                hex
-              }
-            } = message;
-            shouldUseTelemetry() &&
-              reporter.sendTelemetryEvent("colorChanged", {
-                environment,
-                mode,
                 hex,
-                rgb: `${r},${g},${b},${a}`
-              });
+              },
+            } = message;
             this._globalState.update("app", {
               color: message.color,
-              mode: message.mode
+              mode: message.mode,
             });
             const colorString =
               mode === "hex" ? hex : `rgba(${r},${g},${b},${a})`;
             vscode.window.showInformationMessage(
               `${colorString} copied to clipboard`
             );
+            break;
+          }
         }
       },
       null,
@@ -166,18 +110,9 @@ class ReactPanel {
     );
   }
 
-  public doRefactor() {
-    // Send a message to the webview webview.
-    // You can send any JSON serializable data.
-    this._panel.webview.postMessage({ command: "refactor" });
-  }
-
   public dispose() {
     ReactPanel.currentPanel = undefined;
-
-    // Clean up our resources
     this._panel.dispose();
-
     while (this._disposables.length) {
       const x = this._disposables.pop();
       if (x) {
@@ -186,46 +121,49 @@ class ReactPanel {
     }
   }
 
-  private _getHtmlForWebview(initialState: {}) {
-    const manifest = require(path.join(
+  private _getHtmlForWebview(initialState: object) {
+    const webviewPath = path.join(
       this._extensionPath,
       "build",
-      "asset-manifest.json"
-    ));
-    const mainScript = manifest["main.js"];
-
-    const scriptPathOnDisk = vscode.Uri.file(
-      path.join(this._extensionPath, "build", mainScript)
+      "webview"
     );
-    const scriptUri = scriptPathOnDisk.with({ scheme: "vscode-resource" });
+    const manifestPath = path.join(webviewPath, ".vite", "manifest.json");
+    const manifest = JSON.parse(fs.readFileSync(manifestPath, "utf-8"));
 
-    // Use a nonce to whitelist which scripts can be run
+    const entry = manifest["index.html"];
+    const scriptUri = this._panel.webview.asWebviewUri(
+      vscode.Uri.file(path.join(webviewPath, entry.file))
+    );
+
+    const cssLinks = (entry.css || [])
+      .map((cssFile: string) => {
+        const cssUri = this._panel.webview.asWebviewUri(
+          vscode.Uri.file(path.join(webviewPath, cssFile))
+        );
+        return `<link rel="stylesheet" href="${cssUri}">`;
+      })
+      .join("\n");
+
     const nonce = getNonce();
 
-    shouldUseTelemetry() &&
-      reporter.sendTelemetryEvent("extensionStarted", { environment });
-
     return `<!DOCTYPE html>
-			<html lang="en">
-			<head>
-				<meta charset="utf-8">
-				<meta name="viewport" content="width=device-width,initial-scale=1,shrink-to-fit=no">
-				<meta name="theme-color" content="#000000">
-				<title>Color Pick</title>
-				<meta http-equiv="Content-Security-Policy" content="default-src 'none'; img-src vscode-resource: https:; script-src 'nonce-${nonce}';style-src vscode-resource: 'unsafe-inline' http: https: data:;">
-				<base href="${vscode.Uri.file(path.join(this._extensionPath, "build")).with({
-          scheme: "vscode-resource"
-        })}/">
-			</head>
-			<body>
-				<noscript>You need to enable JavaScript to run this app.</noscript>
+      <html lang="en">
+      <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width,initial-scale=1,shrink-to-fit=no">
+        <meta http-equiv="Content-Security-Policy" content="default-src 'none'; img-src ${this._panel.webview.cspSource} https:; script-src 'nonce-${nonce}'; style-src ${this._panel.webview.cspSource} 'unsafe-inline' http: https: data:;">
+        <title>Color Pick</title>
+        ${cssLinks}
+      </head>
+      <body>
+        <noscript>You need to enable JavaScript to run this app.</noscript>
         <div id="root"></div>
-        <script nonce="${nonce}" >
-          INITIAL_COLOR_PICKER_DATA = ${JSON.stringify(initialState)}
-        </script> 
-				<script nonce="${nonce}" src="${scriptUri}"></script>
-			</body>
-			</html>`;
+        <script nonce="${nonce}">
+          window.INITIAL_COLOR_PICKER_DATA = ${JSON.stringify(initialState)};
+        </script>
+        <script nonce="${nonce}" type="module" src="${scriptUri}"></script>
+      </body>
+      </html>`;
   }
 }
 
