@@ -1,10 +1,8 @@
-import * as React from 'react';
-
-interface IIconProps {
-  repoLink: string
+interface IconProps {
+  repoLink: string;
 }
 
-export default (props: IIconProps) => (
+const GithubIcon = (props: IconProps) => (
   <svg id="Layer_1" data-name="Layer 1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 32.58 31.77">
     <title>{props.repoLink}</title>
     <path style={{
@@ -13,3 +11,5 @@ export default (props: IIconProps) => (
     }} d="M152.61,383.44a16.29,16.29,0,0,0-5.15,31.75,0.82,0.82,0,0,0,1.11-.79c0-.39,0-1.41,0-2.77-4.53,1-5.49-2.18-5.49-2.18a4.31,4.31,0,0,0-1.81-2.38c-1.48-1,.11-1,0.11-1a3.42,3.42,0,0,1,2.5,1.68,3.47,3.47,0,0,0,4.74,1.35,3.48,3.48,0,0,1,1-2.18c-3.62-.41-7.42-1.81-7.42-8.05a6.3,6.3,0,0,1,1.68-4.37,5.86,5.86,0,0,1,.16-4.31s1.37-.44,4.48,1.67a15.44,15.44,0,0,1,8.16,0c3.11-2.11,4.48-1.67,4.48-1.67a5.85,5.85,0,0,1,.16,4.31,6.29,6.29,0,0,1,1.67,4.37c0,6.26-3.81,7.63-7.44,8a3.89,3.89,0,0,1,1.11,3c0,2.18,0,3.93,0,4.47a0.82,0.82,0,0,0,1.12.78A16.29,16.29,0,0,0,152.61,383.44Z" transform="translate(-136.32 -383.44)"/>
     </svg>
 );
+
+export default GithubIcon;
